+++ vite.config.js
@@ -17,7 +17,6 @@ export default defineConfig({
       '/api': {
         target: 'http://localhost:8080', // Spring 서버
         changeOrigin: true,
-        
       },
     },
   },
